pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgrss"
version = "0.1.0"
description = ""
authors = ["MasonChance <mason.fryberger@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
django-environ = "^0.4.5"
djangorestframework = "^3.12.2"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
